requirements.txt
# Pipecat framework
pipecat-ai[deepgram,openai,cartesia,webrtc]>=0.0.108

# ASGI server for WebSocket
uvicorn[standard]>=0.30.0
websockets>=13.0

# HTTP server
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
fastapi>=0.115.0

# Configuration and Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

pytz
anthropic
